requirements.txt
# Core interface libraries
typer>=0.9.0
questionary>=2.0.0
rich>=13.0.0
prompt_toolkit>=3.0.39

# Configuration management
PyYAML>=6.0.1

# Development tools (optional)
black>=23.3.0  # Code formatting
pytest>=7.3.1  # Testing
